session, signin

+++ src/main/java/com/pcwk/ehr/member/domain/Member.java
@@ -4,7 +4,7 @@ import com.pcwk.ehr.cmn.DTO;
 
 public class Member extends DTO {
     private String memberId;
-    private int locCode;
+    private long locCode;
     private String password;
     private String name;
     private String nickname;
@@ -39,7 +39,7 @@ public class Member extends DTO {
 		this.memberId = memberId;
 	}
 
-	public int getLocCode() {
+	public long getLocCode() {
 		return locCode;
 	}
